Organize login component
