+++ web/admin.html
@@ -15,7 +15,7 @@
       window.location.href = "index.html";
     }
   </script>
-  <!-- Login UI moved to components/login.html -->
+  <!-- Login UI moved to components/login/login.html -->
   <div id="particles-js" style="position: fixed; width: 100%; height: 100%; z-index: -1; top: 0; left: 0;"></div>
   <div id="particles-js-bg"></div> <!-- Div for animated gradient background -->
   <div id="main-app-content">
