+++ web/components/login/login-main.js
@@ -1,6 +1,6 @@
 import { initLogin } from './login.js';
-import { initParticles } from './particles-config.js';
-import { initIcons } from './icons.js';
+import { initParticles } from '../particles-config.js';
+import { initIcons } from '../icons.js';
 
 document.addEventListener('DOMContentLoaded', () => {
   initLogin();
