+++ web/index.html
@@ -7,10 +7,10 @@
   <link href="https://fonts.googleapis.com/css2?family=Inter:wght@400;700&display=swap" rel="stylesheet">
   <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet">
   <link rel="stylesheet" href="components/style.css">
-  <link rel="stylesheet" href="components/login.css">
+  <link rel="stylesheet" href="components/login/login.css">
 </head>
 <body>
-  <!-- Login UI moved to components/login.html -->
+  <!-- Login UI moved to components/login/login.html -->
   <div id="particles-js" style="position: fixed; width: 100%; height: 100%; z-index: -1; top: 0; left: 0;"></div>
   <div id="particles-js-bg"></div> <!-- Div for animated gradient background -->
 
@@ -19,6 +19,6 @@
   <script src="particles.js"></script>
   <script src="vanilla-tilt.min.js"></script>
   <script src="lucide-icons.js"></script>
-  <script type="module" src="components/login-main.js"></script>
+  <script type="module" src="components/login/login-main.js"></script>
 </body>
 </html>
